skill_scope_ai/backend/requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
pydantic-settings
httpx
python-multipart
pypdf
ollama


